webapi/webapi-webrtc-w3c-tests/webrtc/blink/RTCPeerConnection-localDescription.html

<!DOCTYPE html>
<!--
Test converted from blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/mediastream/RTCPeerConnection-localDescription.html
-->

<meta charset=utf-8>
<title>WebRTC Test: RTCPeerConnection localDescription</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel=help href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#widl-RTCPeerConnection-setLocalDescription-void-RTCSessionDescription-description-VoidFunction-successCallback-RTCPeerConnectionErrorCallback-failureCallback">
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>

var t = async_test(document.title);

var pc = null;
var pc2 = null;
var pc1_offer;

function requestSucceeded1 (offer) {
  pc1_offer = offer;
  t.step(function() {
    pc.setLocalDescription(offer, requestSucceeded2, requestFailed2);
  });
}

function requestFailed1 () {
  t.step(function () {
    assert_unreached("requestFailed1 was called");
  });
  t.done();
}

function requestSucceeded2() {
  pc2 = new webkitRTCPeerConnection(null, null);
  pc2.setRemoteDescription(pc1_offer, requestSucceeded3, requestFailed3);
}

function requestFailed2() {
  t.step(function () {
    assert_unreached("requestFailed2 was called");
  });
  t.done();
}

function requestSucceeded3 () {
  pc2.createAnswer(requestSucceeded4, requestFailed4);
}

function requestFailed3 () {
    t.step(function () {
      assert_unreached("requestFailed3 was called");
    });
    t.done();
}

function requestSucceeded4 (answer) {
  t.step(function() {
    pc2.setLocalDescription(answer, t.step_func(function (ev) {
      t.done();
    }), requestFailed5);
  });
}

function requestFailed4 () {
  t.step(function () {
    assert_unreached("requestFailed4 was called");
  });
  t.done();
}

function requestFailed5() {
  t.step(function () {
    assert_equals(pc.localDescription.type, "offer");
    assert_equals(pc.localDescription.sdp, "local");
    pc.close();
    assert_equals(pc.localDescription.type, "offer");
    assert_equals(pc.localDescription.sdp, "local");
  });
  t.done();
}

t.step(function() {
  pc = new webkitRTCPeerConnection(null, null);
  pc.createOffer(requestSucceeded1, requestFailed1);
});

</script>
